Move historical emission graph to separete component
