+++ app/javascript/app/components/sectors-agriculture/drivers-of-emissions/drivers-of-emissions-styles.scss
@@ -33,21 +33,3 @@
 .switchSelected {
   background-color: $theme-color;
 }
-
-.landscapeContent {
-  @include columns((8,4));
-
-  margin-bottom: 20px;
-
-  &:last-child {
-    height: 520px;
-  }
-}
-
-.portraitContent > * {
-  margin-bottom: 40px;
-}
-
-.dropdown {
-  max-width: 300px;
-}
