+++ app/javascript/app/components/sectors-agriculture/drivers-of-emissions/explore-group/explore-group-component.jsx
