+++ app/javascript/app/components/sectors-agriculture/drivers-of-emissions/historical-emissions-graph/historical-emissions-graph-styles.scss
@@ -1,0 +1,20 @@
+@import '~styles/settings.scss';
+@import '~styles/layout.scss';
+
+.landscapeContent {
+  @include columns((8,4));
+
+  margin-bottom: 20px;
+
+  &:last-child {
+    height: 520px;
+  }
+}
+
+.portraitContent > * {
+  margin-bottom: 40px;
+}
+
+.dropdown {
+  max-width: 300px;
+}
