+++ app/javascript/app/pages/sectors-agriculture/sectors-agriculture-styles.scss
@@ -7,3 +7,7 @@
 .headerContent {
   margin-bottom: 55px;
 }
+
+.sectionComponent {
+  margin-bottom: 100px;
+}
